+++ app.js
@@ -1,4 +1,4 @@
-const testData = require('./data-from-test-work.json');
+const testData = require('./dataFromTestWork');
 
 const helpers = require(
   './1-lesson-homework/helpers/index'
